+++ src/main/java/fr/maxlego08/essentials/commands/CommandLoader.java
@@ -73,6 +73,7 @@ import fr.maxlego08.essentials.commands.commands.utils.CommandHat;
 import fr.maxlego08.essentials.commands.commands.utils.CommandHeal;
 import fr.maxlego08.essentials.commands.commands.utils.CommandMore;
 import fr.maxlego08.essentials.commands.commands.utils.CommandNear;
+import fr.maxlego08.essentials.commands.commands.utils.CommandNightVision;
 import fr.maxlego08.essentials.commands.commands.utils.CommandPlayTime;
 import fr.maxlego08.essentials.commands.commands.utils.CommandRepair;
 import fr.maxlego08.essentials.commands.commands.utils.CommandRepairAll;
@@ -225,6 +226,7 @@ public class CommandLoader {
         register("seen", CommandSeen.class, "whois");
         register("seenip", CommandSeenIp.class, "whoisip");
         register("enchant", CommandEnchant.class, "enchantment");
+        register("nightvision", CommandNightVision.class, "nv");
 
         register("kit", CommandKit.class, "kits");
         register("showkit", CommandShowKit.class);